+++ app/build.gradle
@@ -30,6 +30,10 @@ dependencies {
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+    androidTestCompile 'com.google.dexmaker:dexmaker-mockito:1.0'
+    androidTestCompile 'com.google.dexmaker:dexmaker:1.0'
+    androidTestCompile "org.mockito:mockito-core:1.+"
+
     testCompile 'junit:junit:4.12'
 
     compile 'com.umaplay.oss:fluxxan:1.0.0'
